+++ src/__fixtures__/data/cmsCollections.ts
