+++ src/__tests__/pages/beta/sites-and-applications.test.tsx
@@ -9,21 +9,16 @@ import axios from 'axios'
 
 import { renderWithAuth } from '../../../testHelpers'
 
-import { getCollectionsMock } from '../../../fixtures/getCollection'
-import {
-  sitesAndAppsMock,
-  collectionAdded,
-  collectionsAdded,
-  bookmarkAdded,
-} from '../../../fixtures/sitesAndApps'
-import { cmsBookmarksMock } from '../../../fixtures/cmsBookmarks'
-import { cmsCollectionsMock } from '../../../fixtures/cmsCollections'
+import { getCollectionsMock } from '../../../__fixtures__/operations/getCollection'
+import { cmsBookmarksMock } from '../../../__fixtures__/data/cmsBookmarks'
+import { cmsCollectionsMock } from '../../../__fixtures__/data/cmsCollections'
 import { GET_COLLECTIONS } from 'operations/queries/getCollections'
-
+import { ADD_COLLECTION } from 'operations/mutations/addCollection'
+import { ADD_COLLECTIONS } from 'operations/mutations/addCollections'
+import { ADD_BOOKMARK } from 'operations/mutations/addBookmark'
 import SitesAndApplications, {
   getStaticProps,
 } from 'pages/beta/sites-and-applications'
-import { ADD_BOOKMARK } from 'operations/mutations/addBookmark'
 
 jest.mock('axios')
 
@@ -54,6 +49,99 @@ mockedUseRouter.mockReturnValue({
   asPath: '',
   push: mockPush,
 })
+
+let collectionAdded = false
+let bookmarkAdded = false
+let collectionsAdded = false
+
+const sitesAndAppsMock = [
+  ...getCollectionsMock,
+  {
+    request: {
+      query: ADD_COLLECTION,
+      variables: {
+        title: '',
+        bookmarks: [
+          {
+            cmsId: cmsBookmarksMock[0].id,
+            url: cmsBookmarksMock[0].url,
+            label: cmsBookmarksMock[0].label,
+          },
+        ],
+      },
+    },
+    result: () => {
+      collectionAdded = true
+
+      return {
+        data: {
+          addCollection: {
+            _id: '100',
+            title: '',
+            bookmarks: [
+              {
+                _id: '101',
+                cmsId: cmsBookmarksMock[0].id,
+                url: cmsBookmarksMock[0].url,
+                label: cmsBookmarksMock[0].label,
+              },
+            ],
+          },
+        },
+      }
+    },
+  },
+  {
+    request: {
+      query: ADD_COLLECTIONS,
+      variables: {
+        collections: cmsCollectionsMock,
+      },
+    },
+    result: () => {
+      collectionsAdded = true
+      return {
+        data: {
+          addCollections: cmsCollectionsMock.map((c) => ({
+            _id: '100' + c.id,
+            title: c.title,
+            bookmarks: c.bookmarks.map((b) => ({
+              _id: '101' + b.id,
+              cmsId: b.id,
+              url: b.url,
+              label: b.label,
+            })),
+          })),
+        },
+      }
+    },
+  },
+  {
+    request: {
+      query: ADD_BOOKMARK,
+      variables: {
+        collectionId: getCollectionsMock[0].result.data.collections[0]._id,
+        cmsId: cmsBookmarksMock[0].id,
+        url: cmsBookmarksMock[0].url,
+        label: cmsBookmarksMock[0].label,
+        id: cmsBookmarksMock[0].id,
+      },
+    },
+    result: () => {
+      bookmarkAdded = true
+      return {
+        data: {
+          addBookmark: {
+            _id: '101',
+            cmsId: cmsBookmarksMock[0].id,
+            url: cmsBookmarksMock[0].url,
+            label: cmsBookmarksMock[0].label,
+          },
+        },
+      }
+    },
+  },
+]
 
 describe('Sites and Applications page', () => {
   describe('without a user', () => {
