+++ src/components/MySpace/MySpace.test.tsx
@@ -8,8 +8,8 @@ import React from 'react'
 import { axe } from 'jest-axe'
 import { MockedProvider } from '@apollo/client/testing'
 import { renderWithModalRoot } from '../../testHelpers'
-import { getCollectionsMock } from '../../fixtures/getCollection'
-import { cmsCollectionsMock } from '../../fixtures/cmsCollections'
+import { getCollectionsMock } from '../../__fixtures__/operations/getCollection'
+import { cmsCollectionsMock } from '../../__fixtures__/data/cmsCollections'
 import MySpace from './MySpace'
 import { GET_COLLECTIONS } from 'operations/queries/getCollections'
 import { REMOVE_BOOKMARK } from 'operations/mutations/removeBookmark'
